requirements.txt
numpy==2.2.4
matplotlib==3.10.5
opencv-python==4.11.0.86
pillow==10.4.0
pyyaml==6.0.2
requests==2.32.3
scipy==1.15.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
psutil==7.0.0
pandas==2.2.3
PyQt5==5.15.11